+++ src/crypto/dtls_srtp.c
@@ -367,7 +367,7 @@ static int ms_dtls_srtp_rtp_sendData (void *ctx, const unsigned char *data, size
 	rtp_session_get_transports(session,&rtpt,NULL);
 
 	/* generate message from raw data */
-	msg = rtp_session_create_packet_raw((uint8_t *)data, (int)length);
+	msg = rtp_session_create_packet_raw((uint8_t *)data, length);
 
 	ret = meta_rtp_transport_modifier_inject_packet(rtpt, context->rtp_modifier, msg , 0);
 
@@ -402,7 +402,7 @@ static int ms_dtls_srtp_rtcp_sendData (void *ctx, const unsigned char *data, siz
 	rtp_session_get_transports(session,NULL,&rtcpt);
 
 	/* generate message from raw data */
-	msg = rtp_session_create_packet_raw((uint8_t *)data, (int)length);
+	msg = rtp_session_create_packet_raw((uint8_t *)data, length);
 
 	ret = meta_rtp_transport_modifier_inject_packet(rtcpt, context->rtcp_modifier, msg , 0);
 	freemsg(msg);
